+++ wdl-lint/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 #### Changed
+* Moved command section mixed indentation warnings to `wdl-analysis` and expanded the `MixedIndentation` rule in `wdl-lint` to emit notes for mixed indentation throughout the document, excluding command sections ([#201](https://github.com/stjude-rust-labs/wdl/issues/201)).
 
 * `util::is_properly_quoted` is now `util::is_quote_balanced` ([#341](https://github.com/stjude-rust-labs/wdl/pull/341)).
 * `ShellCheck` is now based on type analysis and is no longer in "beta" ([#341](https://github.com/stjude-rust-labs/wdl/pull/341)).